mmp_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
gunicorn>=21.2
dj-database-url>=1.0
python-dotenv>=1.0
django-celery-beat>=2.5
django-cors-headers==3.14
Pillow>=10.0
requests>=2.31.0